mvp/new_UI/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
psycopg2-binary>=2.9.9
PyJWT>=2.8.0
werkzeug>=3.0.1
PyPDF2>=3.0.1
python-dotenv>=1.0.0
openai>=1.0.0
# Optional: Email digests
sendgrid>=6.10.0
APScheduler>=3.10.0
